+++ artemis/src/main/java/com/artemis/injection/AspectFieldResolver.java
@@ -1,9 +1,15 @@
 package com.artemis.injection;
 
 import com.artemis.*;
+import com.artemis.annotations.AspectDescriptor;
+import com.artemis.utils.reflect.Annotation;
+import com.artemis.utils.reflect.ClassReflection;
 import com.artemis.utils.reflect.Field;
 
 import java.util.IdentityHashMap;
+
+import static com.artemis.Aspect.all;
+import static com.artemis.utils.reflect.ClassReflection.getDeclaredAnnotation;
 
 /**
  * <p>Resolves the following aspect-related types:</p>
@@ -31,17 +37,38 @@ public class AspectFieldResolver implements FieldResolver {
 	@SuppressWarnings("unchecked")
 	public Object resolve(Class<?> fieldType, Field field) {
 		Aspect.Builder aspect = aspect(field);
+		if (aspect == null)
+			return null;
 
 		Class type = field.getType();
 		if (Aspect.class == type) {
-//			world.getAspectSubscriptionManager()
+			return world.getAspectSubscriptionManager().get(aspect).getAspect();
+		} else if (Aspect.Builder.class == type) {
+			return aspect;
+		} else if (EntityTransmuter.class == type) {
+			return new EntityTransmuter(world, aspect);
+		} else if (EntitySubscription.class == type) {
+			return world.getAspectSubscriptionManager().get(aspect);
 		}
 
 		return null;
 	}
 
 	private Aspect.Builder aspect(Field field) {
+		if (!fields.containsKey(field)) {
+			Annotation anno	= field.getDeclaredAnnotation(AspectDescriptor.class);
 
-		return fields.get(field);
+			if (anno != null) {
+				fields.put(field, toAspect(anno.getAnnotation(AspectDescriptor.class)));
+			} else {
+				fields.put(field, null);
+			}
+		}
+
+		return  fields.get(field);
+	}
+
+	private Aspect.Builder toAspect(AspectDescriptor ad) {
+		return all(ad.all()).one(ad.one()).exclude(ad.exclude());
 	}
 }