+++ CHANGELOG.md
@@ -4,8 +4,14 @@
 - **BREAKING CHANGES**
   - World's systems array is now typed with `BaseSystem`; this should only have
     implication for custom implementations of `SystemInvocationStrategy`.
+
+- AspectFieldResolver; additional default injection field resolver
+  - `@AspectDescriptor` on appropriate fields to inject
+  - Valid targets: `Aspect`, `Aspect.Builder`, `EntityTransmuter`, `EntitySubscription`
+- add `EntityTransmuter(World, Aspect.Builder)` constructor
 - **Fix**: WorldConfigurationBuilder.Priority didn't compare properly.
 - **Fix**: thread safefty; InjectionCache's fields no longer static.
+
 
 #### Version: 2.0.0-RC2 - 2016-05-26
 - **BREAKING CHANGES**